category book

+++ database/migrations/2020_10_07_194617_create_book_category_table.php
@@ -15,8 +15,8 @@ class CreateBookCategoryTable extends Migration
     {
         Schema::create('book_category', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('book_id');
-            $table->bigIncrements('category_id');
+            $table->integer('book_id');
+            $table->integer('category_id');
             $table->timestamps();
         });
     }
