+++ database/migrations/2020_10_07_194757_create_order_table.php
@@ -15,7 +15,7 @@ class CreateOrderTable extends Migration
     {
         Schema::create('order', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('user_id');
+            $table->integer('user_id');
             $table->integer('count');
             $table->timestamps();
         });
